+++ demo/solidjs-app/scoped-components.d.ts
@@ -1,8 +1,8 @@
-import type { ScopedElements } from "lit-app/solid-js";
+import type { CustomElements, ScopedElements } from "lit-app/solid-js";
 
 declare module "solid-js" {
   namespace JSX {
     interface IntrinsicElements
-      extends ScopedElements<'test-'>, ScopedElements<'', '_test'> {}
+      extends CustomElements, ScopedElements<'test-'>, ScopedElements<'', '_test'> {}
   }
 }