+++ packages/solidjs-integration/src/type-generator.ts
@@ -1,5 +1,6 @@
 import {
   Component,
+  EXCLUDED_TYPES,
   getComponentDetailsTemplate,
   getComponentProperties,
   getComponents,
@@ -24,25 +25,49 @@ function getOptions(options: Options) {
   return options;
 }
 
+function getEventTypes(component: Component, componentNames: string[]) {
+  const types = component.events
+    ?.map((e) => {
+      const eventType = e.type?.text.replace("[]", "").replace(" | undefined", "");
+      return eventType &&
+        !EXCLUDED_TYPES.includes(eventType) &&
+        !componentNames.includes(eventType) &&
+        !eventType.includes("<") &&
+        !eventType.includes(`{`) &&
+        !eventType.includes("'") &&
+        !eventType.includes(`"`)
+        ? eventType
+        : undefined;
+    })
+    .filter((e) => e !== undefined && !e?.startsWith("HTML"));
+
+  return types?.length ? [...new Set(types)].join(", ") : undefined;
+}
+
 function getTypeTemplate(components: Component[], options: Options) {
+  const componentNames = components
+    .filter((x) => x.customElement)
+    .map((c) => c.name);
   const componentImportStatements =
     typeof options.componentTypePath === "function"
-      ? components.map(
-          (c) =>
-            `import type { ${c.name} } from "${options.componentTypePath!(
-              c.name,
-              c.tagName
-            )}";`
-        )
+      ? components.map((c) => {
+          const types = getEventTypes(c, componentNames);
+          return `import type { ${c.name} ${
+            types ? `, ${types}` : ""
+          } } from "${options.componentTypePath?.(c.name, c.tagName)}";`;
+        })
       : [];
 
   return `
 import type { JSX } from "solid-js";
 ${
   options.globalTypePath
-    ? `import type { ${components.map((c) => c.name).join(", ")} } from "${
-        options.globalTypePath
-      }";`
+    ? `import type { ${components
+        .map((c) => {
+          const types = getEventTypes(c, componentNames);
+          return c.name + (types ? `, ${types}` : "");
+        })
+        .join(", ")} } from "${options.globalTypePath}";`
     : ""
 }
 ${componentImportStatements.join("\n")}
@@ -132,7 +157,6 @@ ${
         event.description,
         event.deprecated
       )} */
-  // @ts-ignore
   "on:${event.name}"?: (e: CustomEvent<${
         event.type?.text || "never"
       }>) => void;`;