+++ packages/solidjs-integration/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.1.1
+
+- Remove duplicate types
+- Add docs for event typing
+
+## 1.1.0
+
+- Added better type resolver for events
+
 ## 1.0.2
 
 - Added deprecation messages to the component description